+++ packages/outline-core/index.ts
@@ -6,7 +6,6 @@ export { LinkedBlockController } from './src/controllers/linked-block-controller
 export { MobileController } from './src/controllers/mobile-controller';
 export { ScreenSizeController } from './src/controllers/screen-size-controller';
 export { SlotsController } from './src/controllers/slots-controller';
-export { LightDomStyles } from './src/controllers/light-dom-styles';
 export { animateTo } from './src/internal/animate';
 export { emit, waitForEvent } from './src/internal/event';
 export { watch } from './src/internal/watch';